Se aplico libreria sweetAlert, se applico asicronias y promesa, semodificaron algunas funciones para intentar un mejor funcionamiento del juego en cuanto a los efectos en los elementos, tambien se creo el paquete JSON

+++ js/script.js
@@ -48,6 +48,7 @@ const message = document.createElement("p");
 const ganar = document.createElement("p");
 const data = JSON.parse(localStorage.getItem('userTurn'));
 
+
 class usuario{
     constructor(nombre, casa){
         this.nombre=nombre;
@@ -57,19 +58,33 @@ class usuario{
     }
 }
 
+////////*Funcion para pedir y guardar informacionde jugador////////
+
 const userData = () => {
 
     message.remove();
+    ganar.remove();
     
     const form = document.querySelector('#form');
     const formContainer = document.querySelector('.form-container-inputs');
 
     const getUserData = () =>{
-        const userName = document.querySelector('#name').value;
-        const casaH = document.querySelector('#casa').value;
-        const usuario1 = new usuario (userName, casaH);
-        localStorage.setItem('userTurn', JSON.stringify(usuario1));  
-    }
+        const userName = document.querySelector('#name').value || "falsy";
+        const casaH = document.querySelector('#casa').value || "falsy";
+        if (userName === "falsy" || casaH === "falsy"){
+            Swal.fire({
+                title: "Debes completar todos los datos para poder continuar",
+                icon: "warning",
+                confirmButtonText: "Aceptar",
+                background: "goldenrod",
+                imageUrl: "https://i.pinimg.com/originals/2c/5d/3f/2c5d3f5dd10f6ca3b7c1f114e6509539.gif"
+                })
+                userData();
+            }else{
+                const usuario1 = new usuario (userName, casaH);
+                localStorage.setItem('userTurn', JSON.stringify(usuario1));
+            } 
+        }
         form.innerHTML = `
         <div>
             <label for="name">Ingresa tu nombre:</label>
@@ -81,7 +96,7 @@ const userData = () => {
         </div>
         <button class="btn" id="save-button">Empezar</button>
         `;
-    
+
         const saveButton = document.getElementById("save-button");
         saveButton.addEventListener("click", getUserData);
 
@@ -90,6 +105,7 @@ const userData = () => {
         return form;
 }
 
+/////////*Contenedor de los atributos de una carta*/////////
 const Card = (prop) => {
 
     let { id, thumbnail} = prop;
@@ -97,15 +113,19 @@ const Card = (prop) => {
     divcard.className= "card-content";
     divcard.innerHTML =
         `
-        <a data-id=${id} class="card-link"><img src=${thumbnail} alt=""></img></a>
+        <a data-id=${id} class="card-link"><img id="card-img" src=${thumbnail} alt="ficha"></img></a>
     `;
 
     const selectedCard = divcard.querySelector('.card-link');
-    selectedCard.addEventListener('click', () => selection(prop));
+    selectedCard.addEventListener('click', () => {
+        divcard.className = "fadeOut";
+        selection(prop);
+    });
 
     return divcard;
 }
 
+/////////*Tablero de cartas*/////////
 const tableContainer = (props) => {
 
     let { cartas } = props;
@@ -118,20 +138,45 @@ const tableContainer = (props) => {
     return section;
 }
 
+/*Informacion del usuario en pantalla*/
 const userTable = ()=>{
     const userInfo = document.createElement("div");
     userInfo.className="user-table";
+
     userInfo.innerHTML = `
     <p class = "welcome">Bienvenido ${data.nombre} de la casa ${data.casa}</p>
-    <p class = "welcome">Pares encontrados: ${found}</p>
+    <p class = "welcome" id="count">Pares encontrados: ${found}</p>
     <div class ="hints">
         <p>Hints: ${data.hints}</p>
         <button class="btn-hints" id="hints-button">Usar</button>
     </div>`;
 
+    const hintsInfo = userInfo.querySelector ("#hints-button");
+    hintsInfo.addEventListener("click", ()=>{
+        Swal.fire({
+            title: "Pistas",
+            icon: "warning",
+            background: "goldenrod",
+            html: `<ul>
+                        <li>
+                            <p>Hipotetica opcion 1</p>
+                        </li>
+                        <li>
+                            <p>Hiotetica opcion 2</p>
+                        </li>
+                    </ul>`,
+            showCancelButton: true,
+            cancelButtonText: "Cancelar",
+            confirmButtonText: "Aceptar",
+        })
+    })
+
     div.appendChild(userInfo);
+
+    return userInfo;
 }
 
+////////*Funcion para orden aleatorio de cartas*/////////
 function shuffle(array) {
     for (let i = array.length - 1; i > 0; i--) {
       const j = Math.floor(Math.random() * (i + 1));
@@ -139,79 +184,115 @@ function shuffle(array) {
     }
 }
 
+////////*Comparo*////////
 function comparar (){
-
+    console.log('comparando')
     const [card1, card2] = cardUp;
-    console.log(cardUp);
     if (card1.id === card2.id) {
-        
         posicion.push(card1);
         posicion.unshift(card2);
         cardUp = [];
-        found=found+1;
-        console.log(found);
-        if (found === cards.length / 2) {
+        found++;
+        console.log("encontre par " + found);
+    }else{
+        console.log('no son iguales');
+        setTimeout(() =>{
+            const newDiscover = document.querySelector(".table-container");
+            const discover = Array.from(newDiscover.children);
+        
+            discover.forEach(element1 => {
+            if(discover.indexOf(element1)===cards.indexOf(card1)){
+                element1.className = "fadeIn";
+                element1.innerHTML =`<a data-id=${card1.id} class="card-link"><img id="card-img" src=${card1.thumbnail} alt="ficha"></img></a>`;
+                const selectCard = element1.querySelector('.card-link');
+                selectCard.addEventListener('click', () => {
+                    element1.className = "fadeOut";
+                    selection(card1);
+                });
+            }
+        });
+        discover.forEach(element2 => {
+            if(discover.indexOf(element2)===cards.indexOf(card2)){
+                element2.className = "fadeIn";
+                element2.innerHTML = `<a data-id=${card2.id} class="card-link"><img id="card-img" src=${card2.thumbnail} alt="ficha"></img></a>`
+                const selectCard = element2.querySelector('.card-link');
+                selectCard.addEventListener('click', () => {
+                    element2.className = "fadeOut";
+                    selection(card2);
+                });
+            }
+        });}, 1000);
+    }
+    
+    if (found === cards.length / 2) {
+        localStorage.clear();
+        const clean1 = document.querySelector(".table-container");
+        const clean2 = document.querySelector(".user-table");
+        clean1.remove();
+        clean2.remove();
+        ganar.className="alerta";
+        ganar.innerText="Ganaste!"
+        div.appendChild(ganar);
+        
+        setTimeout(()=>{
             localStorage.clear();
-            const clean1= document.querySelector(".table-container");
-            const clean2= document.querySelector(".user-table");
-            ganar.className="alerta";
-            ganar.innerText="Ganaste! Haz click para volver a jugar"
-            div.appendChild(ganar);
-            clean1.remove();
-            clean2.remove();
-
-            ganar.addEventListener("click", initGame);
-            
-            
-        }
+            ganar.remove();
+            message.className="start";
+            message.innerText = "Haz click para empezar a jugar";
+            div.appendChild(message);
+            message.addEventListener("click", userData);},2000);
     }else{
         tableContainer({cartas:cards});
         cardUp = [];
     }
-    
 }
 
+/////////*Funcion que se ejecuta al seleccionar o hacer "click" en una carta*///////
 function selection(cart){
     const newDiscover = document.querySelector(".table-container");
     const discover = Array.from(newDiscover.children);
-    message.className ="content";
-    message.innerText = cart.id;
-    cards.forEach(element => {
-        if (element === cart){
-            dato= element;
+    discover.forEach(element1 => {
+        if(cards.indexOf(cart)===discover.indexOf(element1)){
+            element1.className = "fadeIn";
+            element1.innerHTML = `<div id="temp-div"> <p> ${cart.id} </p> </div>`;
         }
-    });
-
-    const select= () => {
         if (cardUp.length < 2 && !cardUp.includes(cart) && !posicion.includes(cart)) {
             cardUp.push(cart);
-            discover.forEach(element1 => {
-                if(cards.indexOf(dato)===discover.indexOf(element1)){
-                    element1.appendChild(message);
-                }
-            });
+            setTimeout(()=>{cardUp.length === 2 && comparar();},1000);
         }
-    }
-    if (cardUp.length === 2) {
-        comparar();
-    }else{
-        select();
-    }
+        
+    });
 }
 
 function initGame(){
+
     message.className="start";
     message.innerText = "Haz click para empezar a jugar";
     div.appendChild(message);
     message.addEventListener("click", userData);
-
-    if (data){
-        message.remove();
-        ganar.remove();
-        shuffle(cards);
-        userTable();
-        div.appendChild(tableContainer({cartas:cards}));
+    const player = (res) => {
+        return new Promise ((resolve, reject) => {
+            if(res){
+                resolve ('Data');
+            }else {
+                reject ('Falta player info');
+            }
+        });
     }
+
+    player (data)
+        .then ((respuesta) => {
+            console.log(respuesta);
+            message.remove();
+            ganar.remove();
+            shuffle(cards);
+            userTable();
+            div.appendChild(tableContainer({cartas:cards}));
+        })
+        .catch((respuesta) => {
+            console.log(respuesta);
+        })
 } 
 
 initGame();
+
